fixes decrypt test
